+++ NebulaCompatibilityAssist/src/Patches/AssemblerVerticalConstruction.cs
@@ -1,4 +1,5 @@
-﻿using crecheng.DSPModSave;
+﻿/*
+using crecheng.DSPModSave;
 using HarmonyLib;
 using NebulaAPI;
 using NebulaCompatibilityAssist.Packets;
@@ -190,3 +191,4 @@ namespace NebulaCompatibilityAssist.Patches
         }
     }
 }
+*/